feat: Add GUI for chatbot
